TASK: Extract `isSuitableFor` out of link type to be testable without mocking
mocking modules as with `jest.mock(` is not that simply possible as it seems.

to allow to use native node js testing all specifications were extracted into simple files to not require react or react-dom at last which would otherwise yield:

> Cannot find package 'react-dom'
