+++ Neos.Ui/core/src/application/LinkTypes/Asset/AssetSpecification.test.js
@@ -1,13 +1,7 @@
-jest.mock('@neos-project/neos-ui-backend-connector', () => ({}), { virtual: true });
-jest.mock('@neos-project/react-ui-components', () => ({}));
-jest.mock('@neos-project/neos-ui-redux-store', () => ({}), { virtual: true });
-jest.mock('@neos-project/neos-ui-editors', () => ({}), { virtual: true });
+import {describe, it} from 'node:test';
+import {equal} from 'node:assert/strict';
 
-jest.mock('../../../presentation', () => ({}));
-
-import '@testing-library/jest-dom/extend-expect';
-
-import {Asset} from './Asset';
+import {isSuitableFor} from './AssetSpecification';
 
 describe('LinkType: Asset', () => {
     it('is not satisfied by http:// links', () => {
@@ -15,8 +9,7 @@ describe('LinkType: Asset', () => {
             href: 'http://www.example.com'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by https:// links', () => {
@@ -24,8 +17,7 @@ describe('LinkType: Asset', () => {
             href: 'https://www.example.com'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by node:// links', () => {
@@ -33,8 +25,7 @@ describe('LinkType: Asset', () => {
             href: 'node://97c9a6e3-4b50-4559-9f60-b5ad68f25758'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is satisfied by asset:// links', () => {
@@ -42,8 +33,7 @@ describe('LinkType: Asset', () => {
             href: 'asset://97c9a6e3-4b50-4559-9f60-b5ad68f25758'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(true);
+        equal(isSuitableFor(link), true);
     });
 
     it('is not satisfied by asset:// links with a hash', () => {
@@ -51,8 +41,7 @@ describe('LinkType: Asset', () => {
             href: 'asset://97c9a6e3-4b50-4559-9f60-b5ad68f25758#section'
         };
 
-        expect(Asset.isSuitableFor(link))
-                .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by mailto: links', () => {
@@ -60,8 +49,7 @@ describe('LinkType: Asset', () => {
             href: 'mailto:foo@example.com'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by invalid links', () => {
@@ -69,8 +57,7 @@ describe('LinkType: Asset', () => {
             href: 'Think of Beethoven\'s 5th: foo foo foo bar'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by tel: links', () => {
@@ -78,7 +65,6 @@ describe('LinkType: Asset', () => {
             href: 'tel:+491258795857'
         };
 
-        expect(Asset.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 });
