+++ Neos.Ui/core/src/application/LinkTypes/MailTo/MailToSpecification.test.js
@@ -1,14 +1,7 @@
-jest.mock('@neos-project/neos-ui-backend-connector', () => ({}), { virtual: true });
-jest.mock('@neos-project/react-ui-components', () => ({}));
-jest.mock('@neos-project/neos-ui-redux-store', () => ({}), { virtual: true });
-jest.mock('@neos-project/neos-ui-editors', () => ({}), { virtual: true });
+import {describe, it} from 'node:test';
+import {equal} from 'node:assert/strict';
 
-jest.mock('../../../presentation', () => ({}));
-jest.mock('../../../framework', () => ({}));
-
-import '@testing-library/jest-dom/extend-expect';
-
-import {MailTo} from './MailTo';
+import {isSuitableFor} from './MailToSpecification';
 
 describe('LinkType: MailTo', () => {
     it('is not satisfied by http:// links', () => {
@@ -16,8 +9,7 @@ describe('LinkType: MailTo', () => {
             href: 'http://www.example.com'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by https:// links', () => {
@@ -25,8 +17,7 @@ describe('LinkType: MailTo', () => {
             href: 'https://www.example.com'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by node:// links', () => {
@@ -34,8 +25,7 @@ describe('LinkType: MailTo', () => {
             href: 'node://97c9a6e3-4b50-4559-9f60-b5ad68f25758'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by asset:// links', () => {
@@ -43,8 +33,7 @@ describe('LinkType: MailTo', () => {
             href: 'asset://97c9a6e3-4b50-4559-9f60-b5ad68f25758'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is satisfied by mailto: links', () => {
@@ -52,8 +41,7 @@ describe('LinkType: MailTo', () => {
             href: 'mailto:foo@example.com'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(true);
+        equal(isSuitableFor(link), true);
     });
 
     it('is not satisfied by invalid links', () => {
@@ -61,8 +49,7 @@ describe('LinkType: MailTo', () => {
             href: 'Think of Beethoven\'s 5th: foo foo foo bar'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 
     it('is not satisfied by tel: links', () => {
@@ -70,7 +57,6 @@ describe('LinkType: MailTo', () => {
             href: 'tel:+491258795857'
         };
 
-        expect(MailTo.isSuitableFor(link))
-            .toBe(false);
+        equal(isSuitableFor(link), false);
     });
 });
