+++ Neos.Ui/core/src/application/LinkTypes/MailTo/MailToSpecification.ts
@@ -1,0 +1,3 @@
+import {ILink} from "../../../domain";
+
+export const isSuitableFor = (link: ILink) => link.href.startsWith('mailto:');
