+++ Neos.Ui/core/src/application/LinkTypes/Node/NodeSpecification.ts
@@ -1,0 +1,3 @@
+import {ILink} from "../../../domain";
+
+export const isSuitableFor = (link: ILink) => link.href.startsWith("node://") && !link.href.includes('#');
