+++ Neos.Ui/core/src/application/LinkTypes/PhoneNumber/PhoneNumberSpecification.ts
@@ -1,0 +1,3 @@
+import {ILink} from "../../../domain";
+
+export const isSuitableFor = (link: ILink) => link.href.startsWith('tel:');
