+++ Neos.Ui/core/src/application/LinkTypes/Web/WebSpecification.ts
@@ -1,0 +1,8 @@
+import {ILink} from "../../../domain";
+
+export const isSuitableFor = (link: ILink) => {
+    const isHttp = link.href.startsWith('http://');
+    const isHttps = link.href.startsWith('https://');
+
+    return isHttp || isHttps;
+};
